deploymentHeroku/iscf_lab1/requirements.txt
psycopg2-binary==2.9.3
flask-restful==0.3.9
Flask-Caching==1.10.1
requests==2.27.1
dash==2.3.1
dash-core-components==2.0.0
dash-html-components==2.0.0
pandas==1.3.5
gunicorn==20.1.0